refactor: extract general chat routing policy
